requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.12,<4.0
psycopg2-binary>=2.9,<3.0
django-cors-headers>=3.10,<4.0
djangorestframework-simplejwt>=5.0,<6.0
pytest>=6.2,<7.0
pytest-django>=4.5,<5.0
django-environ>=0.8,<1.0
drf-spectacular>=0.28,<1.0
django-filter>=24,<30
